requirements.txt
fastapi[standard]
pymongo
python-dotenv
passlib[bcrypt]
pyjwt